refactor: geometryモジュールを削除し、曲線計算機能をcurvesモジュールに移動
