+++ analysis/src/curves.rs
@@ -1,6 +1,6 @@
-//! 幾何特化数値計算モジュール
+//! 曲線・NURBS数値計算モジュール
 //!
-//! NURBS曲線、弧長計算など、幾何形状に特化した数値計算関数を提供する。
+//! NURBS曲線、弧長計算など、曲線形状に特化した数値計算関数を提供する。
 //! 元々model/src/analysis/numeric.rsにあった関数群を独立化。
 use crate::DERIVATIVE_ZERO_THRESHOLD;
